requirements.txt
absl-py==0.7.1
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
colorlover==0.3.0
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.20.0
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipython==7.6.1
ipython-genutils==0.2.0
jedi==0.14.0
jsonschema==3.0.1
jupyter-core==4.5.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
nbformat==4.4.0
networkx==2.3
numpy==1.16.3
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plotly==3.10.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.3
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
requests==2.22.0
retrying==1.3.3
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorboardX==1.6
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.2
wrapt==1.11.2
